# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Digital HubX API.",
        "title": "Digital HubX API",
        "contact": {
            "name": "API Support (Tanapong R)",
            "url": "mailto:zTanapongR@pttep.com",
            "email": "zTanapongR@pttep.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/hubx-service",
    "paths": {
...
        "/admin/question": {
            "post": {
                "description": "Allows an admin to create a question for a class.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Question"
                ],
                "summary": "Create a new question",
                "parameters": [
                    {
                        "description": "Create Question Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateQuestionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question created successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateQuestionRes"
                        }
                    },
                    "400": {
                        "description": "Invalid input\" example:{\"message\":\"Invalid input\",\"status\":400,\"details\":{\"field\":\"error description\"}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error\" example:{\"message\":\"Internal Server Error\",\"status\":500,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/sub-question": {
            "post": {
                "description": "Allows an admin to create a sub-question for a specific choice.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/SubQuestion"
                ],
                "summary": "Create a new sub-question",
                "parameters": [
                    {
                        "description": "Create SubQuestion Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSubQuestionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sub-question created successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSubQuestionRes"
                        }
                    },
                    "400": {
                        "description": "Invalid input\" example:{\"message\":\"Invalid input\",\"status\":400,\"details\":{\"field\":\"error description\"}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error\" example:{\"message\":\"Internal Server Error\",\"status\":500,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/sub-question-choice": {
            "post": {
                "description": "Allows an admin to create a choice for a specific sub-question.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Sub Question Choice"
                ],
                "summary": "Create a new sub-question choice",
                "parameters": [
                    {
                        "description": "Create SubQuestion Choice Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSubQuestionChoicesReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sub-question choice created successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSubQuestionChoicesRes"
                        }
                    },
                    "400": {
                        "description": "Invalid input\" example:{\"message\":\"Invalid input\",\"status\":400,\"details\":{\"field\":\"error description\"}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error\" example:{\"message\":\"Internal Server Error\",\"status\":500,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/question/{class_id}/class": {
            "get": {
                "description": "Retrieves a list of questions associated with a specific class.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Question"
                ],
                "summary": "Get questions by class ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Class ID",
                        "name": "class_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of questions\" example:{\"data\":[],\"total\":0,\"page\":1,\"limit\":10,\"totalPages\":1}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error\" example:{\"message\":\"Internal Server Error\",\"status\":500,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/sub-question/{choice_id}/choice": {
            "get": {
                "description": "Retrieves sub-questions for a given choice ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SubQuestion"
                ],
                "summary": "Get sub-questions by choice ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Choice ID",
                        "name": "choice_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of sub-questions\" example:{\"data\":[],\"total\":0,\"page\":1,\"limit\":10,\"totalPages\":1}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error\" example:{\"message\":\"Internal Server Error\",\"status\":500,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/sub-question/{question_id}/question": {
            "get": {
                "description": "Retrieves sub-questions for a given question ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SubQuestion"
                ],
                "summary": "Get sub-questions by question ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Question ID",
                        "name": "question_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of sub-questions\" example:{\"data\":[],\"total\":0,\"page\":1,\"limit\":10,\"totalPages\":1}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error\" example:{\"message\":\"Internal Server Error\",\"status\":500,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.ChoiceRes": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateQuestionReq": {
            "description": "Request body for creating a question.",
            "type": "object",
            "required": [
                "class_id",
                "description",
                "question_type"
            ],
            "properties": {
                "class_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "question_type": {
                    "$ref": "#/definitions/models.QuestionTypes"
                }
            }
        },
        "dto.CreateQuestionRes": {
            "description": "Response body for creating a question.",
            "type": "object",
            "required": [
                "choices"
            ],
            "properties": {
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Choice"
                    }
                },
                "class": {
                    "$ref": "#/definitions/dto.QuestionClass"
                },
                "class_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "question_type": {
                    "$ref": "#/definitions/models.QuestionTypes"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.CreateSubQuestionChoiceRes": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                }
            }
        },
        "dto.CreateSubQuestionChoicesReq": {
            "description": "Request body for creating a sub-question choice.",
            "type": "object",
            "required": [
                "description",
                "sub_question_id"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "sub_question_id": {
                    "type": "string"
                }
            }
        },
        "dto.CreateSubQuestionChoicesRes": {
            "description": "Response body for creating a sub-question choice.",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "sub_question_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.CreateSubQuestionReq": {
            "description": "Request body for creating a sub-question.",
            "type": "object",
            "required": [
                "choice_id",
                "description",
                "question_type"
            ],
            "properties": {
                "choice_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "question_id": {
                    "description": "? User can ignore these",
                    "type": "string"
                },
                "question_type": {
                    "$ref": "#/definitions/models.QuestionTypes"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.CreateSubQuestionRes": {
            "description": "Response body for creating a sub-question.",
            "type": "object",
            "properties": {
                "choice": {
                    "$ref": "#/definitions/dto.ChoiceRes"
                },
                "choice_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "question": {
                    "$ref": "#/definitions/dto.QuestionRes"
                },
                "question_id": {
                    "type": "string"
                },
                "question_type": {
                    "$ref": "#/definitions/models.QuestionTypes"
                },
                "sub_question_choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CreateSubQuestionChoiceRes"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.QuestionClass": {
            "type": "object",
            "properties": {
                "class_level": {
                    "type": "integer"
                },
                "class_tier": {
                    "$ref": "#/definitions/models.ClassTier"
                },
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.QuestionRes": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                }
            }
        },
...
        "models.Choice": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "question": {
                    "$ref": "#/definitions/models.Question"
                },
                "question_id": {
                    "description": "?",
                    "type": "string"
                },
                "sub_questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SubQuestion"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "user_question_answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserQuestionAnswer"
                    }
                }
            }
        },
...
        "models.Question": {
            "type": "object",
            "properties": {
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Choice"
                    }
                },
                "class": {
                    "$ref": "#/definitions/models.Class"
                },
                "class_id": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "question_type": {
                    "$ref": "#/definitions/models.QuestionTypes"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user_question_answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserQuestionAnswer"
                    }
                }
            }
        },
        "models.QuestionTypes": {
            "type": "string",
            "enum": [
                "choice",
                "text"
            ],
            "x-enum-varnames": [
                "QChoice",
                "QText"
            ]
        },
...
        "models.SubQuestion": {
            "type": "object",
            "properties": {
                "choice": {
                    "$ref": "#/definitions/models.Choice"
                },
                "choice_id": {
                    "description": "เพิ่ม",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "description": "เพิ่ม",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "question": {
                    "$ref": "#/definitions/models.Question"
                },
                "question_id": {
                    "description": "เพิ่ม",
                    "type": "string"
                },
                "question_type": {
                    "$ref": "#/definitions/models.QuestionTypes"
                },
                "sub_question_choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SubQuestionChoice"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "user_sub_question_answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserSubQuestionAnswer"
                    }
                }
            }
        },
...
        "models.UserQuestionAnswer": {
            "type": "object",
            "properties": {
                "answer_text": {
                    "type": "string"
                },
                "choice": {
                    "$ref": "#/definitions/models.Choice"
                },
                "choice_id": {
                    "type": "string"
                },
                "class": {
                    "description": "User     User     `gorm:\"foreignKey:UserEmail;constraint:OnUpdate:CASCADE,OnDelete:CASCADE;\" json:\"user\"`",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Class"
                        }
                    ]
                },
                "class_id": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "question": {
                    "$ref": "#/definitions/models.Question"
                },
                "question_id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user_email": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "This security definition is used for authenticating",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}